src/data/usecases/survey-result/save-survey-result/db-save-survey-result.spec.ts
import { SaveSurveyResultRepository } from './db-save-survey-result-protocols'

import MockDate from 'mockdate'
import { DbSaveResult } from './db-save-survey-result'
import {
  throwError,
  mockSaveSurveyResultParams,
  mockSurveyResultModel,
} from '@/domain/test'
import {
  mockLoadSurveyResultRepository,
  mockSaveResultRepository,
} from '@/data/test'
import { LoadSurveyRepository } from '@/data/protocols/db/surveyResult/load-survey-result-repository'
type SutTypes = {
  sut: DbSaveResult
  saveSurveyResultRepositoryStub: SaveSurveyResultRepository
  loadSurveyResultRepositoryStub: LoadSurveyRepository
}

const makeSut = (): SutTypes => {
  const saveSurveyResultRepositoryStub = mockSaveResultRepository()
  const loadSurveyResultRepositoryStub = mockLoadSurveyResultRepository()
  const sut = new DbSaveResult(
    saveSurveyResultRepositoryStub,
    loadSurveyResultRepositoryStub,
  )

  return {
    saveSurveyResultRepositoryStub,
    sut,
    loadSurveyResultRepositoryStub,
  }
}

describe('DbAddSurvey UseCase', () => {
  beforeAll(() => {
    MockDate.set(new Date())
  })

  afterAll(() => {
    MockDate.reset()
  })

  it('should call saveSurveyResultRepository with correct values', async () => {
    const { sut, saveSurveyResultRepositoryStub } = makeSut()

    const saveSpy = jest.spyOn(saveSurveyResultRepositoryStub, 'save')

    await sut.save(mockSaveSurveyResultParams())

    expect(saveSpy).toHaveBeenCalledWith(mockSaveSurveyResultParams())
  })

  it('should throw if saveSurveyResultRepository throws', async () => {
    const { sut, saveSurveyResultRepositoryStub } = makeSut()

    jest
      .spyOn(saveSurveyResultRepositoryStub, 'save')
      .mockImplementationOnce(throwError)
    const promise = sut.save(mockSaveSurveyResultParams())

    await expect(promise).rejects.toThrow()
  })

  it('should call LoadSurveyResultRepository with correct values', async () => {
    const { sut, loadSurveyResultRepositoryStub } = makeSut()

    const loadSpy = jest.spyOn(loadSurveyResultRepositoryStub, 'loadBySurveyId')

    await sut.save(mockSaveSurveyResultParams())

    expect(loadSpy).toHaveBeenCalledWith(mockSaveSurveyResultParams().surveyId)
  })

  it('should throw if LoadSurveyResultRepository throws', async () => {
    const { sut, loadSurveyResultRepositoryStub } = makeSut()

    jest
      .spyOn(loadSurveyResultRepositoryStub, 'loadBySurveyId')
      .mockImplementationOnce(throwError)
    const promise = sut.save(mockSaveSurveyResultParams())

    await expect(promise).rejects.toThrow()
  })

  it('should return a surveyResult on success', async () => {
    const { sut } = makeSut()

    const surveyResult = await sut.save(mockSaveSurveyResultParams())

    expect(surveyResult).toEqual(mockSurveyResultModel())
  })
})
